requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.7.4
sentence-transformers==2.7.0
transformers==4.41.2
torch==2.3.1
pyPDF2==3.0.1
httpx==0.27.0
